requirements.txt
Babel==2.3.4
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1.1
django-extra-views==0.6.4
django-haystack==2.5.1
django-modeltranslation==0.12
django-oscar==1.4
django-tables2==1.4.2
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.8.1
Faker==0.7.10
funcsigs==1.0.2
ipaddress==1.0.18
mock==2.0.0
MySQL-python==1.2.5
olefile==0.44
pbr==2.0.0
phonenumbers==8.3.3
Pillow==4.0.0
purl==1.3
pycountry==17.1.8
python-dateutil==2.6.0
pytz==2016.10
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
Unidecode==0.4.20
python-memcached==1.58
